Split desktop mutations and test support modules
