+++ crates/defra-agent-desktop/src/app/tests/live/common/mod.rs
@@ -1,0 +1,17 @@
+use super::*;
+
+mod diagnostics;
+mod runtime;
+mod submissions;
+mod tooling;
+
+#[allow(unused_imports)]
+pub(crate) use diagnostics::{compact_field, describe_live_config_state};
+pub(crate) use runtime::{refreshed_runtime_generation, wait_for_stable_runtime_ready};
+pub(crate) use submissions::{
+    assert_live_deployment_default_config, assert_live_submission_rows,
+    wait_for_two_requests_in_flight,
+};
+pub(crate) use tooling::{
+    assert_response_contains_tokens, tool_loop_prompt, wait_for_session_tool_activity,
+};
