+++ crates/defra-agent-desktop/src/app/tests/support/mock_backend/mod.rs
@@ -1,0 +1,17 @@
+use super::*;
+
+mod agent;
+mod backend;
+mod endpoint;
+
+#[allow(unused_imports)]
+pub(crate) use agent::{spawn_backed_agent, wait_for_runtime_process_state, RunningAgent};
+#[allow(unused_imports)]
+pub(crate) use backend::{
+    bind_default_behavior_backend, shutdown_core, test_runtime, AgentBackendConfig,
+};
+#[allow(unused_imports)]
+pub(crate) use endpoint::{
+    extract_desktop_tool_token, mock_completion_sse, mock_tool_call_sse,
+    request_has_tool_result_message, MockModelEndpoint, MockModelMode,
+};
