+++ crates/defra-agent-desktop/src/app/tests/support/response_wait/mod.rs
@@ -1,0 +1,15 @@
+use super::*;
+
+mod diagnostics;
+mod submit;
+
+#[allow(unused_imports)]
+pub(crate) use diagnostics::{describe_response_wait_state, optional_str};
+#[allow(unused_imports)]
+pub(crate) use submit::{
+    submit_chat_message_and_wait_for_observed_response,
+    submit_chat_message_and_wait_for_observed_response_after_request,
+    submit_chat_message_and_wait_for_request_observed, submit_chat_message_and_wait_for_response,
+    submit_chat_message_and_wait_for_response_after_request,
+    wait_for_observed_response_for_request,
+};
