+++ crates/defra-agent-desktop/src/client/mutations/chat/mod.rs
@@ -1,0 +1,9 @@
+mod binding;
+mod conversation;
+mod request;
+
+pub use conversation::{create_conversation, CreatedConversation};
+pub use request::{retry_request, submit_request, SubmittedRequest};
+
+#[cfg(test)]
+mod tests;
