+++ crates/defra-agent-desktop/src/client/mutations/chat/request.rs
@@ -1,0 +1,201 @@
+use anyhow::{Context, Result};
+use chrono::Utc;
+use defra_agent_protocol::row::AgentRequestRow;
+use defra_node::EmbeddedNode;
+use uuid::Uuid;
+
+use crate::client::store::ClientStore;
+
+use super::super::graphql::{
+    escape_graphql_string, execute_mutation, normalize_optional_string, normalize_required,
+};
+use super::binding::resolve_agent_binding;
+use super::conversation::{upsert_conversation, upsert_session};
+
+const DEFAULT_REQUEST_MAX_RETRIES: u32 = 3;
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct SubmittedRequest {
+    pub request_id: String,
+    pub session_id: String,
+    pub agent_did: String,
+    pub behavior_id: Option<String>,
+}
+
+pub async fn submit_request(
+    node: &EmbeddedNode,
+    store: &ClientStore,
+    session_id: &str,
+    agent_did: &str,
+    content: &str,
+    behavior_id: Option<&str>,
+) -> Result<SubmittedRequest> {
+    let session_id = normalize_required("session_id", session_id)?;
+    let agent_did = normalize_required("agent_did", agent_did)?;
+    let content = normalize_required("content", content)?;
+    let request_id = Uuid::new_v4().to_string();
+    let created_at = Utc::now().to_rfc3339();
+    let binding = resolve_agent_binding(store, agent_did, behavior_id, Some(session_id))?;
+
+    upsert_session(
+        node,
+        store,
+        session_id,
+        &binding.agent_name,
+        &binding.behavior_id,
+    )
+    .await?;
+
+    let escaped_request_id = escape_graphql_string(&request_id);
+    let escaped_agent_did = escape_graphql_string(agent_did);
+    let escaped_behavior_id = escape_graphql_string(binding.behavior_id.as_deref().unwrap_or(""));
+    let escaped_session_id = escape_graphql_string(session_id);
+    let escaped_content = escape_graphql_string(content);
+    let escaped_created_at = escape_graphql_string(&created_at);
+
+    let mutation = format!(
+        r#"mutation {{
+            add_AgentRequest(input: {{
+                request_id: "{escaped_request_id}",
+                agent_did: "{escaped_agent_did}",
+                behavior_id: "{escaped_behavior_id}",
+                session_id: "{escaped_session_id}",
+                retry_parent_request: "",
+                retry_root_request: "{escaped_request_id}",
+                superseded_by_request: "",
+                content: "{escaped_content}",
+                status: "pending",
+                lifecycle_state: "pending",
+                backend_id: "",
+                execution_origin: "interactive",
+                failure_reason: "",
+                created_at: "{escaped_created_at}",
+                retry_count: 0,
+                max_retries: {max_retries}
+            }}) {{ _docID }}
+        }}"#,
+        max_retries = DEFAULT_REQUEST_MAX_RETRIES,
+    );
+    execute_mutation(node, &mutation, "submit_request").await?;
+
+    upsert_conversation(
+        node,
+        store,
+        session_id,
+        agent_did,
+        &binding.agent_name,
+        &binding.behavior_id,
+        &request_id,
+        content,
+        "active",
+    )
+    .await?;
+
+    Ok(SubmittedRequest {
+        request_id,
+        session_id: session_id.to_string(),
+        agent_did: agent_did.to_string(),
+        behavior_id: binding.behavior_id,
+    })
+}
+
+pub async fn retry_request(
+    node: &EmbeddedNode,
+    store: &ClientStore,
+    parent: &AgentRequestRow,
+) -> Result<SubmittedRequest> {
+    let parent_request_id = normalize_required("request_id", &parent.request_id)?;
+    let session_id = normalize_required(
+        "session_id",
+        parent
+            .session_id
+            .as_deref()
+            .context("retry parent request must have a session_id")?,
+    )?;
+    let agent_did = normalize_required(
+        "agent_did",
+        parent
+            .agent_did
+            .as_deref()
+            .context("retry parent request must have an agent_did")?,
+    )?;
+    let content = normalize_required(
+        "content",
+        parent
+            .content
+            .as_deref()
+            .context("retry parent request must have content")?,
+    )?;
+    let behavior_id = normalize_optional_string(parent.behavior_id.as_deref());
+    let retry_root_request = normalize_optional_string(parent.retry_root_request.as_deref())
+        .unwrap_or(parent_request_id);
+    let retry_count = parent.retry_count.unwrap_or_default() + 1;
+    let max_retries = parent
+        .max_retries
+        .unwrap_or(i64::from(DEFAULT_REQUEST_MAX_RETRIES));
+    let request_id = Uuid::new_v4().to_string();
+    let created_at = Utc::now().to_rfc3339();
+    let binding = resolve_agent_binding(store, agent_did, behavior_id, Some(session_id))?;
+
+    upsert_session(
+        node,
+        store,
+        session_id,
+        &binding.agent_name,
+        &binding.behavior_id,
+    )
+    .await?;
+
+    let escaped_request_id = escape_graphql_string(&request_id);
+    let escaped_parent_request_id = escape_graphql_string(parent_request_id);
+    let escaped_retry_root_request = escape_graphql_string(retry_root_request);
+    let escaped_agent_did = escape_graphql_string(agent_did);
+    let escaped_behavior_id = escape_graphql_string(binding.behavior_id.as_deref().unwrap_or(""));
+    let escaped_session_id = escape_graphql_string(session_id);
+    let escaped_content = escape_graphql_string(content);
+    let escaped_created_at = escape_graphql_string(&created_at);
+
+    let mutation = format!(
+        r#"mutation {{
+            add_AgentRequest(input: {{
+                request_id: "{escaped_request_id}",
+                agent_did: "{escaped_agent_did}",
+                behavior_id: "{escaped_behavior_id}",
+                session_id: "{escaped_session_id}",
+                retry_parent_request: "{escaped_parent_request_id}",
+                retry_root_request: "{escaped_retry_root_request}",
+                superseded_by_request: "",
+                content: "{escaped_content}",
+                status: "pending",
+                lifecycle_state: "pending",
+                backend_id: "",
+                execution_origin: "interactive",
+                failure_reason: "",
+                created_at: "{escaped_created_at}",
+                retry_count: {retry_count},
+                max_retries: {max_retries}
+            }}) {{ _docID }}
+        }}"#
+    );
+    execute_mutation(node, &mutation, "retry_request").await?;
+
+    upsert_conversation(
+        node,
+        store,
+        session_id,
+        agent_did,
+        &binding.agent_name,
+        &binding.behavior_id,
+        &request_id,
+        content,
+        "active",
+    )
+    .await?;
+
+    Ok(SubmittedRequest {
+        request_id,
+        session_id: session_id.to_string(),
+        agent_did: agent_did.to_string(),
+        behavior_id: binding.behavior_id,
+    })
+}
