+++ crates/defra-agent-desktop/src/client/mutations/operator/mod.rs
@@ -1,0 +1,11 @@
+mod behavior;
+mod inference;
+mod profile;
+mod scheduled;
+mod tools;
+
+pub use behavior::upsert_agent_behavior;
+pub use inference::upsert_inference_backend;
+pub use profile::upsert_inference_profile;
+pub use scheduled::{run_scheduled_task_now, upsert_scheduled_task};
+pub use tools::upsert_tool_selection;
